+++ packages/bento-design-system/src/defaultTheme.ts
@@ -1,0 +1,1 @@
+import "./defaultTheme.css";
